+++ packages/sdk-client/src/api/api-client-options.ts
@@ -29,6 +29,12 @@ interface BaseApiClientOptions {
    *  @default false
    */
   useServicePlanId?: boolean;
+
+  /**
+   * Request/connection timeout in seconds for HTTP I/O.
+   * @default 60
+   */
+  timeoutSeconds?: number;
 }
 
 /** @internal */